==> CODE/03_analysis.Rmd <==

# Setup Step: Setup libraries
```{r}
# Load necessary libraries
library(data.table)
library(dplyr)
library(zoo)  # for na.locf function
library(lubridate)
library(tidyr)
library(fst)
library(arrow)
library(table1)
library(metafor)
library(yaml)
library(htmltools)
library(here)

config <- yaml::read_yaml(here("config.yaml"))

# Define directories based on YAML configuration
data_dir <- config$data_dir
intermediate_dir <-  file.path(here(), "sedation_language_project", "intermediate_data")  # Define output_dir based on data_dir
output_dir <- file.path(here(), "sedation_language_project", "final_data")
file_type <- config$file_type
institution <- config$institution


# Create the output directory if it does not exist
if (!dir.exists(output_dir)) {
  dir.create(output_dir, recursive = TRUE)  # 'recursive = TRUE' allows creating nested directories
}
```

# PART ONE: Summary Statistics
```{r}
# Upload Part One Data Table
part_two_data <- fread(paste0(intermediate_dir, "analytic_table.csv.gz"))

# Identify reference points for model
part_two_data$language <- relevel(as.factor(part_two_data$language), ref = "English")
part_two_data$ethnicity <- relevel(as.factor(part_two_data$ethnicity), ref = "Non-Hispanic")
part_two_data$race <- relevel(as.factor(part_two_data$race), ref = "White")
part_two_data$sex <- relevel(as.factor(part_two_data$sex), ref = "Male")
part_two_data$primary_sedative <- relevel(as.factor(part_two_data$primary_sedative), ref = "propofol")

# Create labels for readability
label(part_two_data$race) <- "Race"
label(part_two_data$ethnicity) <- "Ethnicity"
label(part_two_data$sex) <- "Sex"
label(part_two_data$age) <- "Age"
label(part_two_data$language) <- "Language"
label(part_two_data$bmi) <- "BMI"
label(part_two_data$paralyzed) <- "Paralyzed"
label(part_two_data$total_sofa) <- "Average SOFA"
label(part_two_data$hospital_id) <- "Hospital"
label(part_two_data$primary_sedative) <- "Primary Sedative"

# Generate the summary table grouped by language
summary_table <- table1(~ age + sex + ethnicity + race + bmi + paralyzed + total_sofa + hospital_id + primary_sedative | language, data = part_two_data)

# Convert to HTML and save as file
save_html(as.tags(summary_table), file = paste0(output_dir, institution, "_total_cohort_language_summary.html"))

```

# PART TWO: Primary Analysis
```{r}

# Fit the unadjusted primary model
primary_model_unadjusted <- lm(percent_deep_sedation ~ language, data = part_two_data)

# Extract coefficients, standard errors, variance, and p-values for the unadjusted model
primary_results_unadjusted <- data.frame(
  term = names(coef(primary_model_unadjusted)),
  estimate = coef(primary_model_unadjusted),
  std.error = summary(primary_model_unadjusted)$coefficients[, "Std. Error"],
  variance = (summary(primary_model_unadjusted)$coefficients[, "Std. Error"])^2,  # Calculate variance
  p.value = summary(primary_model_unadjusted)$coefficients[, "Pr(>|t|)"]  # Extract p-values
)

summary(primary_model_unadjusted)


# Fit the adjusted primary model
n_hospitals <- part_two_data$hospital_id %>% unique() %>% length()

if (n_hospitals>1){
  primary_model_adjusted <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + 
                               race + bmi + paralyzed + total_sofa + 
                               factor(hospital_id) + factor(primary_sedative), 
                             data = part_two_data)
} else {
    primary_model_adjusted <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + 
                               race + bmi + paralyzed + total_sofa + 
                               factor(primary_sedative), 
                             data = part_two_data)
}

summary(primary_model_adjusted)

# Extract coefficients, standard errors, variance, and p-values for the adjusted model
primary_results_adjusted <- data.frame(
  term = names(coef(primary_model_adjusted)),
  estimate = coef(primary_model_adjusted),
  std.error = summary(primary_model_adjusted)$coefficients[, "Std. Error"],
  variance = (summary(primary_model_adjusted)$coefficients[, "Std. Error"])^2,  # Calculate variance
  p.value = summary(primary_model_adjusted)$coefficients[, "Pr(>|t|)"]  # Extract p-values
)

# Combine the results into one data frame
primary_results_combined <- bind_rows(
  primary_results_unadjusted %>% mutate(model = "Unadjusted"),
  primary_results_adjusted %>% mutate(model = "Adjusted")
)

# Save the combined results to a file
fwrite(primary_results_combined, file = paste0(output_dir, institution, "_primary_analysis_results.csv.gz"))

```


# PART THREE: Secondary Analysis
```{r}
# Secondary Model of Total IMV Time
if (n_hospitals>1){
  secondary_model_imv_time <- lm(total_imv_time ~ language + age + sex + ethnicity + race + bmi + paralyzed + total_sofa + factor(hospital_id), data = part_two_data)
} else {
secondary_model_imv_time <- lm(total_imv_time ~ language + age + sex + ethnicity + race + bmi + paralyzed + total_sofa, data = part_two_data)
}
summary(secondary_model_imv_time)


# Extract coefficients and statistics for total imv time model
secondary_results_imv_time <- data.frame(
  term = names(coef(secondary_model_imv_time)),
  estimate = coef(secondary_model_imv_time),
  std.error = summary(secondary_model_imv_time)$coefficients[, "Std. Error"],
  variance = (summary(secondary_model_imv_time)$coefficients[, "Std. Error"])^2,  # Calculate variance
  p.value = summary(secondary_model_imv_time)$coefficients[, "Pr(>|t|)"]  # Extract p-values from the correct model
)



# Save to file
fwrite(secondary_results_imv_time, file = paste0(output_dir, institution, "_secondary_analysis_results_imv_time.csv.gz"))


# Secondary Model of 30-Day Mortality
if (n_hospitals >1 ) {
  secondary_model_mortality <- glm(thirty_day_mortality ~ language + age + sex + ethnicity + race + bmi + paralyzed + total_sofa + factor(hospital_id), 
                                  data = part_two_data, 
                                  family = binomial(link = "logit"))
} else{
  secondary_model_mortality <- glm(thirty_day_mortality ~ language + age + sex + ethnicity + race + bmi + paralyzed + total_sofa, 
                                  data = part_two_data, 
                                  family = binomial(link = "logit"))
  
}

summary(secondary_model_mortality)

# Extract coefficients and calculate odds ratios, standard errors, and variance
coef_summary <- summary(secondary_model_mortality)$coefficients
odds_ratios <- exp(coef_summary[, "Estimate"])  # Exponentiated coefficients
standard_errors <- coef_summary[, "Std. Error"]
variance <- standard_errors^2  # Calculate variance
p_values <- coef_summary[, "Pr(>|z|)"]  # Extract p-values

# Create a data frame to hold the results
secondary_results_df <- data.frame(
    Variable = rownames(coef_summary),
    Odds_Ratio = odds_ratios,
    Std_Error = standard_errors,
    Variance = variance,  # Add variance
    P_Value = p_values    # Add p-values
)

# Write the results to a CSV file
fwrite(secondary_results_df, file = paste0(output_dir, institution, "_secondary_analysis_results_mortality.csv.gz"))

```


# PART FOUR: Sub-Group Analysis
```{r}

extract_model_results <- function(model, model_name) {
  coef_summary <- summary(model)$coefficients
  results_df <- data.frame(
    term = rownames(coef_summary),
    estimate = coef(model),
    std.error = coef_summary[, "Std. Error"],
    variance = (coef_summary[, "Std. Error"])^2,  # Calculate variance
    p.value = coef_summary[, "Pr(>|t|)"]  # Extract p-values for linear models
  )
  results_df$model <- model_name
  return(results_df)
}

# STEP ONE:
# HISPANIC/LATINO ANALYSIS
hispanic_data <- part_two_data[ethnicity == "Hispanic"]
hispanic_data$language <- relevel(as.factor(hispanic_data$language), ref = "English")
hispanic_data$race <- relevel(as.factor(hispanic_data$race), ref = "White")
hispanic_data$sex <- relevel(as.factor(hispanic_data$sex), ref = "Male")
hispanic_data$primary_sedative <- relevel(as.factor(hispanic_data$primary_sedative), ref = "propofol")

# Generate the summary table grouped by language
summary_table_hispanic <- table1(~ age + sex + race + bmi + paralyzed + total_sofa + hospital_id + primary_sedative | language, data = hispanic_data)

# Convert to HTML and save as file
save_html(as.tags(summary_table_hispanic), file = paste0(output_dir, institution, "_hispanic_cohort_language_summary.html"))

# Fit a linear model for Hispanic patients, excluding ethnicity
if (n_hospitals >1){
  model_hispanic <- lm(percent_deep_sedation ~ language + age + sex + race + bmi + paralyzed + total_sofa + factor(hospital_id) + factor(primary_sedative), data = hispanic_data)
} else{
    model_hispanic <- lm(percent_deep_sedation ~ language + age + sex + race + bmi + paralyzed + total_sofa + factor(primary_sedative), data = hispanic_data)
}

summary(model_hispanic)
results_hispanic <- extract_model_results(model_hispanic, "Hispanic")


# STEP TWO:
# BLACK/AFRICAN AMERICAN ANALYSIS
black_data <- part_two_data[race == "Black or African American"]
black_data$language <- relevel(as.factor(black_data$language), ref = "English")
black_data$ethnicity <- relevel(as.factor(black_data$ethnicity), ref = "Non-Hispanic")
black_data$sex <- relevel(as.factor(black_data$sex), ref = "Male")
black_data$primary_sedative <- relevel(as.factor(black_data$primary_sedative), ref = "propofol")

# Generate the summary table grouped by language
summary_table_aa <- table1(~ age + sex + ethnicity + bmi + paralyzed + total_sofa + hospital_id + primary_sedative | language, data = black_data)

# Convert to HTML and save as file
save_html(as.tags(summary_table_aa), file = paste0(output_dir, institution, "_aa_cohort_language_summary.html"))

# Fit a linear model for Black patients, excluding race
if (n_hospitals>1){
  model_black <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + bmi + paralyzed + total_sofa + factor(hospital_id) + factor(primary_sedative), data = black_data)
} else{
    model_black <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + bmi + paralyzed + total_sofa + factor(primary_sedative), data = black_data)
}

summary(model_black)
results_black <- extract_model_results(model_black, "Black or African American")


# STEP THREE:
# ASIAN ANALYSIS
asian_data <- part_two_data[race == "Asian"]
asian_data$language <- relevel(as.factor(asian_data$language), ref = "English")
asian_data$ethnicity <- relevel(as.factor(asian_data$ethnicity), ref = "Non-Hispanic")
asian_data$sex <- relevel(as.factor(asian_data$sex), ref = "Male")
asian_data$primary_sedative <- relevel(as.factor(asian_data$primary_sedative), ref = "propofol")

# Generate the summary table grouped by language
summary_table_asian <- table1(~ age + sex + ethnicity + bmi + paralyzed + total_sofa + hospital_id + primary_sedative | language, data = asian_data)

# Convert to HTML and save as file
save_html(as.tags(summary_table_asian), file = paste0(output_dir, institution, "_asian_cohort_language_summary.html"))

# Fit a linear model for Asian patients
if (n_hospitals >1){
  model_asian <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + bmi + paralyzed + total_sofa + factor(hospital_id) + factor(primary_sedative), data = asian_data)
} else{
    model_asian <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + bmi + paralyzed + total_sofa + factor(primary_sedative), data = asian_data)
}

summary(model_asian)
results_asian <- extract_model_results(model_asian, "Asian")


# STEP FOUR:
# WHITE ANALYSIS
white_data <- part_two_data[race == "White"]
white_data$language <- relevel(as.factor(white_data$language), ref = "English")
white_data$ethnicity <- relevel(as.factor(white_data$ethnicity), ref = "Non-Hispanic")
white_data$sex <- relevel(as.factor(white_data$sex), ref = "Male")
white_data$primary_sedative <- relevel(as.factor(white_data$primary_sedative), ref = "propofol")

# Generate the summary table grouped by language
summary_table_white <- table1(~ age + sex + ethnicity + bmi + paralyzed + total_sofa + hospital_id + primary_sedative | language, data = white_data)

# Convert to HTML and save as file
save_html(as.tags(summary_table_white), file = paste0(output_dir, institution, "_white_cohort_language_summary.html"))


# Fit a linear model for White patients
if (n_hospitals>1){
  model_white <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + bmi + paralyzed + total_sofa + factor(hospital_id) + factor(primary_sedative), data = white_data)
} else{
    model_white <- lm(percent_deep_sedation ~ language + age + sex + ethnicity + bmi + paralyzed + total_sofa + factor(primary_sedative), data = white_data)
}

summary(model_white)
results_white <- extract_model_results(model_white, "White")

# Combine all results into one data frame
combined_subgroup_results <- bind_rows(results_hispanic, results_black, results_asian, results_white)

# Save to file
fwrite(combined_subgroup_results, file = paste0(output_dir, institution, "_subgroup_analysis_results.csv.gz"))

```
